+++ silver_set.sh
@@ -6,7 +6,7 @@
 #SBATCH --partition=scc-gpu
 #SBATCH -G A100:1
 #SBATCH --cpus-per-task=16
-#SBATCH --time=06:00:00
+#SBATCH --time=14:00:00
 
 set -euo pipefail